# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EPS2 Dossier Patient — API",
        "description": "API REST pour la gestion des dossiers patients des établissements publics de santé du Sénégal (EPS2). Authentification via Bearer Token (Laravel Sanctum).",
        "contact": {
            "name": "Support EPS2",
            "email": "support@eps2.sn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/consultations": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Liste des consultations",
                "description": "Retourne les 50 consultations les plus récentes avec les informations du patient associé. **Réservé aux Médecins et Administrateurs.**",
                "operationId": "consultationIndex",
                "responses": {
                    "200": {
                        "description": "Liste des consultations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "allOf": [
                                            {
                                                "$ref": "#/components/schemas/Consultation"
                                            },
                                            {
                                                "properties": {
                                                    "nom": {
                                                        "type": "string",
                                                        "example": "Diallo"
                                                    },
                                                    "prenom": {
                                                        "type": "string",
                                                        "example": "Aminata"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "PAT-0001"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé — rôle Médecin ou Administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Créer une consultation",
                "description": "Enregistre une nouvelle consultation médicale pour un patient. Le médecin est automatiquement renseigné avec l'utilisateur connecté si non fourni. **Réservé aux Médecins et Administrateurs.**",
                "operationId": "consultationStore",
                "requestBody": {
                    "description": "Données de la consultation",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "description": "Identifiant du patient",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "description": "Date de la consultation (aujourd'hui par défaut)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-10"
                                    },
                                    "medecin": {
                                        "description": "Nom du médecin (utilisateur connecté par défaut)",
                                        "type": "string",
                                        "example": "Dr Fatou Ndiaye",
                                        "nullable": true
                                    },
                                    "motif": {
                                        "type": "string",
                                        "example": "Douleur abdominale",
                                        "nullable": true
                                    },
                                    "diagnostic": {
                                        "type": "string",
                                        "example": "Gastrite aiguë",
                                        "nullable": true
                                    },
                                    "traitement": {
                                        "type": "string",
                                        "example": "Oméprazole 20 mg",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consultation enregistrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consultation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Patient requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé — rôle Médecin ou Administrateur requis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Consultation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "patient_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-06-10"
                    },
                    "medecin": {
                        "type": "string",
                        "example": "Dr Fatou Ndiaye"
                    },
                    "motif": {
                        "type": "string",
                        "example": "Douleur abdominale",
                        "nullable": true
                    },
                    "diagnostic": {
                        "type": "string",
                        "example": "Gastrite aiguë",
                        "nullable": true
                    },
                    "traitement": {
                        "type": "string",
                        "example": "Oméprazole 20 mg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Message d'erreur"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Obtenez un token via POST /api/login, puis transmettez-le dans le header : Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}